requirements.txt
gradio
openai-whisper
openai
pytube
moviepy
imageio[ffmpeg]
